requirements.txt
Django==6.0.1
django_ckeditor==6.7.3
pdfplumber==0.11.9
python-decouple==3.8
Requests==2.32.5
pillow == 12.1.0
